% Auto-generated: do not edit by hand
\name{'dsl'SlickSlider}

\alias{'dsl'SlickSlider}

\title{SlickSlider component}

\description{

}

\usage{
'dsl'SlickSlider(children=NULL, id=NULL, arrows=NULL, autoplay=NULL,
center_mode=NULL, center_padding=NULL, className=NULL,
dots=NULL, infinite=NULL, labels=NULL, responsive=NULL,
slide_navigator=NULL, slides_to_scroll=NULL,
slides_to_show=NULL, speed=NULL, style=NULL,
swipe_to_slide=NULL, variable_width=NULL, vertical=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Your carousel is vertical}

\item{id}{Character. Id of the element}

\item{arrows}{Logical. Arrows to control carousel}

\item{autoplay}{Logical. If the carousel will start to play automatically}

\item{center_mode}{Logical. To centralize the carousel}

\item{center_padding}{Character. Padding on the sides}

\item{className}{Character. Style class of the component.}

\item{dots}{Logical. Dots under carousel}

\item{infinite}{Logical. If the carousel content will repeat in a loop}

\item{labels}{List of characters. Your labels for dropdown if slide_navigator is set to True}

\item{responsive}{Unnamed list. Settings of breakpoints}

\item{slide_navigator}{Logical. Render a dropdown that navigates between slides}

\item{slides_to_scroll}{Numeric. How many slides will scroll when you swipe the carousel}

\item{slides_to_show}{Numeric. How many slides you want to show}

\item{speed}{Numeric. Speed of autoplay}

\item{style}{Named list. Inline style of the component.}

\item{swipe_to_slide}{Logical. If you can slide to scroll the slides}

\item{variable_width}{Logical. The slides width varies according to the screen size}

\item{vertical}{Logical. If your carousel is vertical}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

